+++ src/MCAL/RCC/RCC.h
@@ -288,7 +288,7 @@ ErrorStatus_t RCC_ConfigurePLL(PLLCFG_t* PLL_Config);
  * @return: ErrorStatus_t Status of the operation (OK if successful, WRONG_PARAMETER if invalid
  *         parameters are provided).
  */
-ErrorStatus_t RCC_ControlPeripheralCLK(uint32_t *BusRegister,uint8_t Peripheral,uint8_t PeripheralCFG);
+ErrorStatus_t RCC_ControlPeripheralCLK(uint32_t *BusRegister,uint32_t Peripheral,uint32_t PeripheralCFG);
 
 /**
   * @brief : Configures the prescaler for the Low Speed APB1 bus.
